+++ lib/scripts/eventPage.js
@@ -1,7 +1,9 @@
 import {BingTranslate} from './services/bingTranslate'
-import * as contextMenuFeatures from './services/contextMenuFeaturesImplementation'
+import {ContextMenu} from './services/contextMenu'
 
 console.log('eventPage running');
+
+ var contextMenuFeatures = new ContextMenu();
 
 //sets up default data in localStorage
 function setupDefaultData() {
